Added visuals to coffee-making process
I also added speakers in the cafe so that the sound of the radio comes out of them instead of the radio itself. Another thing I did was seperate the stand equipment from the place where the player sits and rests at, so the coffee cups can be placed on top of that surface once they are completed and received.